+++ src/main/java/org/betonquest/betonquest/conversation/ChatConvIO.java
@@ -10,6 +10,7 @@ import org.betonquest.betonquest.util.Utils;
 import org.bukkit.Bukkit;
 import org.bukkit.ChatColor;
 import org.bukkit.Location;
+import org.bukkit.configuration.ConfigurationSection;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.EventPriority;
 import org.bukkit.event.HandlerList;
@@ -164,7 +165,7 @@ public abstract class ChatConvIO implements ConversationIO, Listener {
     }
 
     @Override
-    public void addPlayerOption(final String option) {
+    public void addPlayerOption(final String option, final ConfigurationSection properties) {
         optionsCount++;
         options.put(optionsCount, option);
     }