+++ src/main/java/org/betonquest/betonquest/conversation/ConversationIO.java
@@ -1,6 +1,8 @@
 package org.betonquest.betonquest.conversation;
 
 import net.kyori.adventure.text.Component;
+import org.betonquest.betonquest.api.quest.QuestException;
+import org.bukkit.configuration.ConfigurationSection;
 import org.jetbrains.annotations.Nullable;
 
 /**
@@ -18,12 +20,13 @@ public interface ConversationIO {
     void setNpcResponse(Component npcName, Component response);
 
     /**
-     * Adds the text of the player option. Should be called for each option in a
-     * conversation cycle.
+     * Adds the text of the player option. Should be called for each option in a conversation cycle.
      *
-     * @param option the text of an option
+     * @param option     the text of an option
+     * @param properties the property configuration section for the text
+     * @throws QuestException if the option cannot be added, e.g., if the properties are invalid
      */
-    void addPlayerOption(String option);
+    void addPlayerOption(String option, ConfigurationSection properties) throws QuestException;
 
     /**
      * Displays all data to the player. Should be called after setting all