+++ CHANGELOG.md
@@ -71,6 +71,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
     - `owner:%player%` was changed to `owner:` to allow constant pre-parsing
 - `math` variable now gives a better exception message and does not return 0 instead
 - `folder` event now use the `unit` argument to define the time unit for ticks, seconds and minutes
+- custom items in the chest conversation IO are now defined in the `properties` section on an option
 ### Deprecated
 ### Removed
 - undocumented prefix feature in conversation